% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lavaan.printer-package.R
\docType{package}
\name{lavaan.printer-package}
\alias{lavaan.printer}
\alias{lavaan.printer-package}
\title{lavaan.printer: Helper Functions for Printing 'lavaan' Outputs}
\description{
Helpers for customizing selected outputs from 'lavaan' by Rosseel (2012) \doi{10.18637/jss.v048.i02} and print them. The functions are intended to be used by package developers in their packages and so are not designed to be user-friendly. They are designed to be let developers customize the tables by other functions. Currently the parameter estimates tables of a fitted object are supported.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://sfcheung.github.io/lavaan.printer/}
}

}
\author{
\strong{Maintainer}: Shu Fai Cheung \email{shufai.cheung@gmail.com} (\href{https://orcid.org/0000-0002-9871-9448}{ORCID})

}
\keyword{internal}
